MH-11620 : Fix redirection to `index.html` for unprivileged users
The previous commit to this ticket made it possible for unprivileged
users with the `ROLE_ADMIN_UI` to access the `index.html` page in a
Presentation node. This works if the unprivileged user browses to
`<server_url>/index.html`, but it fails to redirect to this page when
the user visits simply `<server_url>` or `<server_url>/`. In other
words, the redirection does not work for unprivileged users.

This suggestion by Lars Kiesow <lkiesow@uos.de> fixes this problem.