MH-11702 : Set default series creation date when the index is rebuilt
This fixes the issue where series created in older Opencast/Matterhorn
versions do not have a creation date defined. The patch provides a
sensible default (the epoch --1st Jan. 1970--), so that the index is
correctly built, and the series are thus displayed in the UI.

Please note that this patch does not edit the series representation in
the database, so the default creation date is used every time the index
is rebuilt. This is the main reason why a constant default (the Epoch)
and not the current date is used as default --if that was the case, the
default creation date for those series that do not have one would vary
in an unpredictable way (an index rebuild can be triggered at any time).