MH-11739 : Add new 'tags-and-flavors' parameter to Compose Operation
This new parameter changes the default behaviour to match input tracks among
those that have EITHER one of the configured 'source-flavors' OR one of the
'source-tags' and selects tracks that match BOTH any of the 'source-flavors' AND
any of the 'source-tags'.

The use case is to use tags to narrow the tracks selected by flavor. For
instance, select a specific delivery track among multiple copies with different
qualities.

Conflicts:
	modules/matterhorn-conductor/src/main/java/org/opencastproject/workflow/handler/ComposeWorkflowOperationHandler.java