MH-11716: Initial work on getting external roles to appear.
The gist of it is that the role needs to appear in the list generated by the ROLES resource endpoint, but since it's an
external role that was added via the /series rest endpoints it does not appear in this list.  This code pushes the
missing roles into the list of roles on the client side by parsing the list of roles from the series itself.  Angular,
however, has other ideas, and is hitting race conditions.  If you put a breakpoint in the return line for the ROLES to
make sure they occur first, everything works fine, but without that breakpoint it does not.