MH-11750, Mark Annotation Module as Deprecated
As discussed before, the annotation tool shall be removed. At the moment
there might, however, be some Paella player users still using this for
its soft-trimming feature.

The outcome of the discussion about this was to mark the service as
deprecated and remove it in two versions from now. By then it should be
no problem for Paella anymore.

Signed-off-by: Lars Kiesow <lkiesow@uos.de>