MH-11883, Larger files may remain in system temp directory
The Zip workflow operation handler will create zip archives by copying
all files belonging to a media package to a temporary directory and then
zipping this directory. If Opencast is inconveniently shut down, these
temporary files will stay there forever.

Furthermore, the location of the directory is poorly chosen since this
is often not the same storage space used by the workspace. Hence no hard
linking of files can be used which slows down the operation.

This patch ensures that the directory is cleaned during service
activation and a temporary directory is created in the main storage
directory.

Signed-off-by: Lars Kiesow <lkiesow@uos.de>