MH-11889, Row size too large
The automatic creation of the table `mh_organization_property` will
fail. The row size is too large. The maximum row size for the used table
type, not counting BLOBs, is 65535.

The ddl script uses a TEXT field instead. This patch will have the same
effect for the automatic table creation.

Signed-off-by: Lars Kiesow <lkiesow@uos.de>