MH-11962, Missing slidetext.xml should not break re-indexing
This patch catches the error for a missing `slidetext.xml` during
re-indexing of the search Solr index and catch it rather than breaking
the indexing for the whole event.

This still causes the slide texts not being included in the index but at
least the events are listed at all.

--

Cherry-picked for 2.2.5