[BugFix] abandon use of job arguments from jobs created by distribution service
Internal encoding of job arguments has been changed in distribution service (JSON)
which broke publication service code. However since job arguments shall be internal
to a service the correct solution is to not rely on job arguements in client code
of a service. Job and element ID are now being passed as a tuple.