MH-12019, Ensure Test Files Are Deleted
Several of Opencast's unit tests create temporary files in /tmp but may
fail to delete them at the end of the test. Especially, if tests fail.

This patch introduces JUnit's TemporaryFolder as a means of safely
creating temporary files for tests, ensuring that the files or folders
are always removed at the end of the testing.

Signed-off-by: Lars Kiesow <lkiesow@uos.de>