MH-12084 : Fix AsyncTimeoutRedirectFilter not handling exceptions correctly
The AsyncTimeoutRedirectFilter catches all the exceptions that happen in the
filterchain downstream and returns a redirection when it detects two specific
exception types. However, it fails to return any other exception that does not
match these categories. This makes debugging very difficult, because requests
that fail because of an exception do not properly return this failure to the
caller, but a "200 OK" response instead (because that is the response value
when initialized).

This patch makes sure that all the exceptions types that are not relevant to
this filter are rethrown to ensure a proper error response is returned.