MH-12087 Add null check for bitrate
See also MH-11224, MH-11241 and MH-11340. The Entwine functional libraries do not
like null values. This can cause errors when the resulting object is being serialized to
JSON. Because no useful information is logged, it is extremely difficult to track these
down. Values added like this should always use vN if it's possible that the value could
be null.