MH-12013 let database set and update the modification timestamp of OAI-PMH entities
It has been reported that the modification timestamps sometimes differ
very much from the time the OaiPmhEntity is actually inserted into the
database. That may lead to records never being picked up by harvesters
if they harvest in the time span between setting the timestamp and the
entity appearing in the database.