MH-12091 Implement per-tenant digest user for capture agents
NOTE ABOUT 5.x CHERRY-PICK
--------------------------

This patch only picks the asset manager modification in preparation for
the security patch fixing CVE-2018-19899

--------------------------

This PR adds a new in-memory per-tenant user and a new role
ROLE_CAPTURE_AGENT for capture agents. It builds upon work by
@smarquard and @lkiesow, see the bitbucket PR here:

https://bitbucket.org/opencast-community/opencast/pull-requests/1542

The new user is able to digest-auth in order to ingest recordings and
thus has to be an in-memory user with clear-text password. Restricting
the user's rights turned out to be rather tricky because of a
"somewhat circular" dependency between AssetManagerWithMessaging and
AssetManagerWithSecurity.

Co-authored-by: Stephen Marquard <stephen.marquard@uct.ac.za>
Co-authored-by: Lars Kiesow <lkiesow@uos.de>