MH-12083 : Improve handling of security contexts
With this commit, the SecurityService always authenticates a user coming from
an existing SecurityContext, even if the UserDirectory cannot find it.

The behaviour introduced by this commit is as follows:

* If the "delegatedUserHolder" already contains a non-null user, return that
user.
* Otherwise, if a SecurityContext exists:
  * Get the username in the SecurityContext and search it in the UserDirectory
  * If the user exists in the UserDirectory:
    * Return a user with the same username and the combination of the roles
      defined in SecurityContext *AND* in the user returned by the UserDirectory
  * If the user **DOES NOT** exist in the UserDirectory:
    * Return a user with the roles contained in the SecurityContext
* Otherwise, if a SecurityContext does *NOT* exist:
  * Return the anonymous user

Finally, this commit also makes sure that the returned user is previously stored
in the "delegatedUserHolder", so that subsequent calls to this function will
return faster.