Require Java 8
As discussed on list, Opencast > 2.4.x will require Java 8 to be able to
use all the “cool new things”.

This patch updates the main `pom.xml` to require Java 8 and fixes two
tests which started to complain. To make things easier for the future,
the Java version is now specified as one global variable.

This is to be merged into `develop` only after the `r/2.4.x` branch is
created.