MH-12109 comments can be cretaed again.
Comments could only be created when a job was read_only (ng-disable=!transactions.read_only) which is contrary to the delete where the same statement was with ng-show instead ng-disable. In genereal comments seem to be createable anytime, so no need for the restriction here. Probably the backend was broken in the past so that this bug did not show up.