MH-12163 Events with stopped workflows sometimes cannot be deleted
When a workflow is stopped, set the workflow job to CANCELED rather than STOPPED. This
allows workflow information to be retrieved (for example for viewing in the UI)
rather than getting a notfound error.

When removing an event with a stopped workflow, update the stop method to first
check if the state is already stopped, to avoid an unnecessary update call that
can also cause failures if the workflow mediapackage is not fully available.