MH-12180, Specify Value For probe-resolution
This patch fixes an error in my reasoning when coming up with the
probe-resolution workflow operation handler. The idea was for it to be
able to handle special cases the automated analyze-tracks operation would
not cover.

While it does that by introducing new variables, it cannot properly
overwrite analyze-tracks variables since the data type and value format
do not match.

This patch fixes that issue by letting users configure the actual value
that is set if resolutions match. This will allow you to refine the
pre-set values.

As an example, this would automatically detect and set the aspect ratio
but for the special case of 960x720 which is probably 1280x720 with a
non-square pixel aspect ratio which is then covered by probe-resolution.

    <operation
      id="analyze-tracks"
      fail-on-error="true"
      exception-handler-workflow="ng-partial-error"
      description="Analyze tracks in media package and set control variables">
      <configurations>
        <configuration key="source-flavor">*/source</configuration>
        <configuration key="aspect-ratio">4/3,16/9</configuration>
      </configurations>
    </operation>

    <operation
      id="probe-resolution"
      fail-on-error="true"
      exception-handler-workflow="ng-partial-error"
      description="Set control variables based on video resolution">
      <configurations>
        <configuration key="source-flavor">*/source</configuration>
        <configuration key="var-aspect_snap">960x720</configuration>
        <configuration key="val-aspect_snap">16/9</configuration>
      </configurations>
    </operation>

After these two operations you can use the single variable
`…_aspect_snap` to check if the video has, for example, a 16/9 aspect
ratio.