MH-12302, Use Sensible Default Configuration
This patch changes the asset manager default configuration to use the
URL of the admin node as default with another default to the servers own
URL. This should work on both all-in-one and distributed set-ups.

This patch also changes the log level of which URL is actually used to
debug to avoid cluttering the logs with unnecessary warnings.