MH-12271 Update existing files atomically in the working file repository
This is to avoid race conditions when one or more indexing threads are attempting
to read an episode.xml DC catalog file while a workflow operation (such as inspect)
is updating it.

The ATOMIC_MOVE operation is not guaranteed to be available on all filesystems,
so the implementation falls back to REPLACE_EXISTING if necessary.