MH-12318, Remove Webconsole Default Installation
This patch removes the default deployment of the Karaf web console
(/system/console) to avoid security risks caused by this tool which
allows extended control of the application server via a web interface.
This includes installing, starting and stopping of bundles and features
as well as reading and writing bundle and system configuration.

Note that this is only removed by default. The web console can still be
installed as feature using the Karaf Shell by running:

    feature:install webconsole

This will hopefully make security issues due to bugs or
misconfigurations a lot less likely while in the rare cases where the
console is needed, it can still be installed.