MH-12428, Event Metadata Performance Issue
Opencast will try to dynamically translate *all* drop-down fields
regardless if they can actually be translated or not causing massive
performance issues in the dialog when several thousand data sets are
present.

For example, series identifier and titles are presented as drop-down in
the metadata dialog. They can not be translated, but it is constantly
tried. When adding a few thousand series, this makes the add event and
edit event dialog extremely slow.

This patch adds a *translatable* property based on which the admin
interface now decides if data should be translated or not. The
translatability is hard coded for the static list providers, but is
configurable for custom list providers.