MH-12423 Fix Date Parse Error When Chaning Metadata
For metadata fields which use a select field, the id of the changed
field wasn't set correctly which resulted in all fields being sent to
the server. Since the fields are sent in a "raw" manner, the value for
start date is not parsable because its format depends on the users
locale/language setting.