MH-12500: Fix misuse of method 'URL#getFile' to convert a resource URL into a file path
The name of the 'getFile' method is misleading: it returns the URL path plus
the query string concatenated, if there is one. It is not guaranteed that such
a string will be a valid system path, particularly when the file path contains
special characters such as whitespaces.

The function 'URL#toURI' should be used instead. The obtained URI can then be
used to instantiate a 'File' or a 'Path' object.