MH-12448, Remove unconfigured send-mail WOH
The send-mail WOH in the error workflow does not make a lot of sense
with its default configuration since it will send a mail to
`root@localhost`, hence this patch removes it. Users can easily decide
to add it again by taking the example from the documentation with proper
configuration values.