MH-12478, Waveform filenames are not unique
The waveform service uses a video's filename as a basis for the waveform
filename. If two files with the same name (e.g. track.mp4) are processed
simultaneously, the service will overwrite the output files and
operations will either end up with a wrong waveform or the operation
will fail due to one operation already having removed the waveform from
the workspace collection.

This patch fixes that issue by adding the track's identifier to the
filename since the service should not work twice on the same track.