MH-12510: Rewrite some of the ActiveMQ stuff to be more robust.
There are a few problems with the current state of the ActiveMQ connection
module that are fixed by this commit:

1. In `MessageBaseFacility`, there's a boolean value `enabled` that indicates if
there's a connection to ActiveMQ. The `TransportListener` listens for connection
changes and sets this boolean accordingly. This is not done in a thread-safe
way.
2. The bigger problem, however, is that ActiveMQ is not configured to reconnect
automatically on connection loss. At least not indefinitely. Just twice,
actually (see `maxReconnectAttempts`). So internally, ActiveMQ tries to
reconnect twice (in the background) and then gives up. After that, the `enabled`
boolean is never set to true. This is bad, because the custom reconnect code
checks if `enabled` is false. Thus, the connection will never recover.
3. Curiously, there are two ways in which the module checks if a connection
exists: the aforementioned `enabled` boolean and the `isConnected()` function,
which calls `getMetadata` on the connection object (again, not thread-safe).
4. The reconnect logic doesn't reconnect every 60 seconds, as desired, since
`InterruptedExceptions`...interrupt the waiting process.
5. The only method that's `synchronized` is the connect method. All other
functions are not thread-safe.
6. The disconnect function doesn't reset the `enabled` boolean and acts as if
the connection is immediately there.