MH-12511: Deleting with inconsistent search index state doesn't work
This patch adds another REST endpoint "hasSnapshots.json" indicating if an event
has snapshots in the asset manager. If this fails, it's an indication that the
index and the asset manager data base have become inconsistent.

We use this new method when the user clicks on the delete icon in the event
overview. Where previously, one of the dialogs "retract" or "delete" was chosen
using an `ng-if`, we now do it more dynamically, by checking the result of the
REST call.