Make POST tasks/new Asynchronous
When starting tasks via recordings -> actions -> start task, the UI is
blocked until the tasks are created. On productive systems, this can
take up to minutes depending on the performance of the backend and on
the amount of data it holds.

This patch changes the behavior of the TasksEndpoint. Instead of waiting
for the tasks to be created and responding afterwards, an immediate
response is issued while the tasks are being created in the background.
To indicate that the tasks are not yet running, the workflow state of
the corresponding events is set to "PENDING". If any error occurs during
task creation, the workflow state is set to "FAILED".

This way, the "Create Task" wizard is closed immediately and the user
can go on editing other events.