MH-12525 : Allow non-admin users to modify their own event's ACLs
There was a typo in UI and the role `ROLE_UI_EVENT_DETAILS_ACL_EDIT` was used
instead of the correct one `ROLE_UI_EVENTS_DETAILS_ACL_EDIT`.

Additionally, the non-existing variable `aclLocked` was being used in the event
wizard UI. This is likely the result of copying the same fragment of code in the
series wizard, which does use the variable (why two components providing
essentially the same functionality have duplicated, almost identical UIs is
beyond me). In other identical parts of the two UIs, the variable `aclLocked` in
the series UI is substituted by the expression `transactions.read_only` in the
event UI. Therefore, the remaining uses of the `aclLocked` variable have been
modified the same way.