MH-12526 : Return the current user when 'all LDAP users' are requested
Including a way to list all the "Opencast-enabled" users in an LDAP instance is
questionable, because it opens the door to multiple security issues and it does
not scale in the long run.

On the other hand, the implementation of the UserIdRoleProvider is such that it
requests all users from all providers defined in Opencast in order to generate
a list of available users roles in the system. Should the LdapUserProvider not
return at least the current logged-in user, the roles lists displayed in the UI
will not be correctly populated.

This commit implements a compromise solution by returning only the currently
logged-in user, if such user exist in the LDAP directory.

A long-term solution should refactor the UserIdRoleProvider to take a more
scalable approach, but that lies outside of the scope of this bugfix.