MH-12560 : Fixed remote streaming distribution
The commit `39ea6f14c037809a118655147784d21eb4560433` introduced a very small
change in the `StreamingDistributionService`'s API, by modifying the name of a
POST parameter from `elementId` to `elementIds`. An additional commit
(`aba6aa9078715d18941595f023f8c751b07152b2`) brought that change to the
`WowzaStreamingDistributionService`. Such a change was, however, not implemented
in the "remote" implementation of the streaming service, which meant that the
streaming services would only work when running in a all-in-one or an
admin-presentation server.

This commit fixes the bug by updating the POST parameter name in the "remote"
implementation of the streaming service.