Fix First Editable in Wizards Not Being Editable
Due to a race condition, the first editable in the Event and Series
Wizards are not editable when opnening the wizard for the first time
after page reload.

This happens, because the focushere directive causes the first Editable
to enter the edit mode while there is still CSS class manipulation going
on in the DOM tree. This fix avoids the CSS manipulation by using ng-if
instead of ng-show on the affected elements.