MH-12000, Cross-tenant URL signing
Authorised users can use the endpoints /api/security/sign and
/signing/sign to let Opencast sign a URL provided by the user.  Opencast
will determine if a URL signing key has been configured for a prefix of
the provided URL and, if so, return the signed URL that has been signed
with the configured signing key.

The problem is that URL signing keys are not specific to tenants and
therefore, an authorised user can sign URLs of other tenants.

This patch fixes that problem by adding an organization context to these
signing requests allowing users to only sign URLs specific to their
tenants.

Signed-off-by: Lars Kiesow <lkiesow@uos.de>