MH-12577, Do not make content public if no permissions are defined
This patch modifies the default ACL to allow read and write access by
system (and possibly organization) administrators only, prohibiting all
access from anonymous users.

Note that this ACL will only be applied if neither an event ACL, nor a
series ACL is defined which should seldom be the case but may happen is
rare cases.