MH-12594: Implement proper support for newlines in long_text fields
This code adds the ability to hold down "Shift" to add a newline into
"long text" fields like the "Description" field of a series or an
event.

This indirectly fixes a bug whereby pressing enter in the
middle of a long text actually inserted a newline, but didn't display
it in the UI.

Also, the textarea that appears in edit mode is now constrained to be
resized only on the y axis, so the dialog layout cannot be broken
anymore.