MH-12596, Video Editor Ignores Workspace
The video editor will use the local tmp directory to process large video
files which might cause problems when this directory is located in a
small filesystem local to the VM. All other large files processed by
Opencast are put into the workspace instead.

This patch changes this behavior and makes the editor use a sub-folder
of the workspace instead as well.