MH-12476 Add a delay of 60s before starting to dispatch jobs on startup
There are some cases where job dispatching too close to startup can cause operation failures.

One situation is where admin node is behind a proxy (a common configuration),
so the admin node itself is on http://server.domain:8080/ but the proxy public
URL is http://server.domain/

If Opencast restarts, the proxy (e.g. apache) can mark the service as offline
and requests get a service-not-available 503 error for a short time even after
the service has come back, because the proxy has a timeout, as described for
the retry parameter here:

https://httpd.apache.org/docs/current/mod/mod_proxy.html#proxypass

So Opencast checks for available services on http://server.domain/ and gets
back nothing available even though it's up.

So it is safer to add a delay before starting to dispatch jobs on startup.
This is most significant for scenarios where Opencast admin node is restarted
while workflows are running (usually for unplanned reasons), and avoids
workflows failing because no services are available.

The 60s delay value is the default value of ProxyTimeout for apache's mod_proxy.