MH-12638 Migration bundle needs to have a higher runlevel
The migration bundle can start up before the REST handler for the asset manager
is registered in some cases:

(RestPublisher:333) - Registered REST endpoint at /assets

This will cause a migration failure because /assets URLs will get a 404.

The runlevel of the migration bundle needs to be higher than other bundles.