Upgrade angular-chosen & chosen
This updates angular-chosen to version 1.8.0. With this version, two-way
data binding is working. Unfortunately, it is not possible to include
angular-chosen through bower since this pulls in a very recent version
of jquery which breaks lots of other stuff.

You can see the now working two-way data binding in action when you open
the  video-editor and play around with the upper and lower drop-downs
where the workflow can be selected. Before this patch. the drop-downs
used to become out of sync. Now, the model is updated correctly and they
always share the same value.