MH-12660 Scheduling Events by Specifying End Time
This patch adds the possibility to specify the end time when scheduling
events. The behavior is as follows:

- When the start time is changed, the end time is kept and the duration
  is adjusted accordingly.
- When the duration is changed, the start time is kept and the end time
  is adjusted accordingly.
- When the end time is changed, the start time is kept and the duration
  is adjusted accordingly.

The following views are affected by this change:

- Add Event -> Schedule Multiple
- Add Event -> Scheduling Single
- Event Details -> Scheduling

The latter two views have one speciality: When start date + duration
lead to the end date being the next day, the end date is displayed
besides the end time. However, for universities, it should be a rare
use case to schedule events over midnight.