MH-12641 Asset manager conflict checks are very slow
When scheduling new events for a certain capture agent, the
scheduler service checks for conflicts with existing events already
scheduled on that capture agent.

For this check, queries are created using the asset manager's query
builder. Unfortunately, due to the nature of the underlying database
schema, the resulting SQL queries contain many nested queries and hence,
are very slow if there already is a higher number of scheduled events.

The best approach to solve this would be to change the database schema.
The problem is mainly caused by the concept of generic 'properties'
which are related to snapshots and are managed by the asset manager.
However, changing the schema would mean that we need a database
migration and it would yield architectural issues since the separation
between asset manager and scheduler service would break.

Since there currently is not enough time to change the database schema,
this patch chooses a different approach. The measurements include:

- Don't do three queries for each event to schedule, but only one query
  for all the events to schedule. When scheduling 50 events, we now use
  one query instead of 150 queries.
- Use a simpler query with less nested queries.

This is achieved by reading all events for the capture agent in
question and then doing the conflict checks in memory. Of course, there
may be lots of events and performance issues can still arise when
reading so many rows from the database. Also, no limit is used which
generally is bad practice. In the worst case, opencast could run out of
memory or reading all the rows from the database could take forever.
However, productive systems are expected to have a moderate number of
scheduled recordings per capture agent which should not cause the
problems mentioned.

Disclaimer: This solution is not optimal since it utilizes an anti
pattern, but it solves the performance issue for now. It should be
considered as a temporary workaround which as well has its problems.