MH-12664, Ensure series can be deleted
In a few cases, existing series cannot be deleted due to Solr index
problems. The reason is that the index service will check if a series
exists in the index before attempting to delete it which is quite
unnecessary since that check will be done by the series service on the
real data again.

This patch removes the additional check allowing a series to be deleted
even if there is a faulty index entry.