MH-12665 Sort table on startup
If there is no entry in local storage containing the sorting criteria
for the table in the admin ui (e.g. if starting opencast for the first
time in this browser or if local storage has been emptied), simply sort
by first sortable column. This way the table is never unsorted.