MH-12678 Fix Video Playback After Tab Switch
This basically re-introduces the code removed by commit #28e6fc053.

When tabs are switched, the location is updated to end with 'editor' or
'playback', depending on the active tab. However, it turns out that it
is necessary to preserve the route in order to avoid breaking the video
playback when changing the location. Unfortunately, preserving the route
breaks the close button. So the route-restoring code must only be used
when switching tabs, not when closing the video editor/player.

If the route is changed, the <video> element starts to behave odd when
controlled by opencast's external controls. While it still plays audio,
the video image becomes frozen.