MH-12730, Workflow Index Rebuild Performance
For a workflow-index rebuild Opencast loads and deserializes far more
data that it actually needs. This may even lead to a system needing
hours before it can actually start the index rebuild or for the process
to fail due to the massive amount of required memory.

This patch tries to improve this behavior by requesting the specific
resources directly from the database and deserializing only what is
necessary.

Additionally, it will also avoid sending those messages via message
broker which would only be used for creating log messages. Instead, the
logging is executed directly.