MH-12684 Make License List Provider More Flexible
This patch adds a new metadata type 'ordered_text' which allows for
listproviders to configure the order, as well as a "selectable" value
for each entry. This can be used to display e.g. licenses in a
meaningful order as well hiding entries if desired. "Hiding" doesn't
actually mean hiding in this case. If "selectable" is false for an
entry, the entry is hidden as long as it isn't already selected as the
current value.

For licenses, there is the following use case: Previously, Creative
Commons 3.0 was selected by users to publish their recordings. Now,
there is Creative Commons 4.0 added to the list and the list of licenses
grows longer and longer. While it is not feasible to change the license
for old events, for new events only Creative Commons 4.0 should show up
in the list. So we cannot remove Creative Commons 3.0, but can hide it
for new events as well as for old events, which were published under a
different license. This is now possible by setting "selectable" to false
for old licenses.