MH-12675 Ensure consistent date type
Ensure that the params.value always contains an ISO String instead of a
Date Object even if the date is never changed via the date-time-picker.
Otherwise the display of the date sometimes changes after it was edited
(for example 29/01/2018 09:00 to Today at 09:00 if the former is today),
additionally there are errors during processing if the date is not of
the expected type.