MH-12675 Send start date also if it hasn't changed
When creating a new event, send the displayed default start date
(current date and time) to the backend if it hasn't been changed instead
of leaving that metadata field blank.  This patch also fixes the bug
where an additional row will appear at the end of the metadata tab if
the start date was changed in the source tab.