MH-12692: Update maven bundle plugin for java8
This commit updates the maven-bundle-plugin to 3.x, which updates the
“bnd” utility used to resolve osgi dependencies.

By upgrading the plugin, it fixes a bug (a known bug triggering an
ArrayOutOfBoundsException) which prevented Java 8 from being used in
production code. However, it also introduced an ambiguity. Both
“jsr311” and “jsr339” provide the Java package
“javax.rs.ws” (apparently, JSR-331 is JAX-RS 1.1 and JSR-339 is
version 2.0). This lead to an ambiguous import chain, for example in
matterhorn-admin-ui-ng.

The solution employed here is to force OSGI to use version 2 of the
“javax.rs.ws” package (and replacing a dependency in the
feature.xml). All the code using JAX-RS still compiles and runs.

This also fixes a bug in the pom of one of the bundles, where using
${opencast.version} lead to a NumberFormatException.

Also, I couldn’t resist my OCD and fixed one or two indentation
mishaps in the vicinity of the bundle-plugin directives I was
changing.