MH-12713: Do not import `org.snowball*` in `search` and `solr`
Both of these modules embed these packages as part of Lucene,
and both also export them. Both do not import Lucene, though,
so they are guaranteed to use the embedded Lucene classes.
However, these in turn might end up using an incompatible version
of the snowball packages, depending on the order
in which `solr` and `search` are loaded.