MH-12711 Improve errorhandling during xacmlparsing
Get more detailed error messages when parsing of an XACML fails, handle
unexpected errors like for example IndexOutOfBoundsExceptions when the
role is missing. Was initially supposed to fix an error where actions
like viewing the details of an event become impossible when the ACL is
malformed (see issue description), hoewever the root of that problem
most likely lies elsewhere.