MH-12826: Make workflow processing settings persistent
This PR adds “workflow properties” to events, as described here:

https://opencast.jira.com/browse/MH-12826

The idea is to store the last workflow instance’s properties using
asset manager properties and load them when the user invokes “Actions
-> Start Task”. This, of course, has to work when multiple events are
selected, which makes the code a lot more difficult.

For example, if two events are selected, both have the (boolean,
checkbox) workflow property “publish to X”, but with different
values. Then the gui now displays an indeterminate checkbox. In a
similar fashion, radio buttons are all deselected for ambiguous
values, and text fields are blanked.

The “Schedule Task” wizard also has a table at the end listing all the
selected properties. This table was changed to display an asterisk for
properties that are different for each event. When hitting the submit
button, the corresponding workflow properties are resolved for each
event separately.

There’s a new endpoint to retrieve these properties in bulk for
multiple events (for performance reasons). To accomplish that, the
`AssetManager` query language had to be extended.

Co-authored-by: Christian Greweling <christian.greweling@uni-osnabrueck.de>