MH-12786 Undistinguishable Entries in Groups Editor User List
When two users share the same fore- and surname, they cannot be
distinguished from each other in the following lists:

- Add Group -> users: left and right list (available and selected)
- Edit Group -> users: right list (selected)

This patch fixes the issue by also displaying the username, or whatever
is configured by prop.adminui.user.listname in
org.opencastproject.organization.cfg.

Also, the translation filter is removed from the selectContainer since
it causes problems when displaying email addresses. Translating entries
of the selectContaiener seems to not be used by now.