MH-12797, Explain UI Actions (Events)
The events table contains several quick filters and actions. For
example, clicking on a series will automatically add a filter for that
particular series which can be very useful. However, the functionality
of these actions is explained nowhere and users are left to figure that
out by their own.

This patch adds title tags to all undescribed icons and links giving a
very short explanation about the functionality when users hover over
them.