MH-12831, Fixing dependencies
This patch is part of getting Opencast getting to not fail with Java 9+.
It fixes a lot of dependency problems I've found when making the
necessary changes for Java 9+ (e.g. explicitly adding java.ee dependencies).

Note that despite the large number of changes the only two non-test,
non-dependency fixes are two places in our code where `commons.lang` was
still used instead of `commons.lang3`.

The main parts of this patch are:

- Explicitly declaring the java.ee dependencies:
  https://stackoverflow.com/a/43574427/2352895
- Fixing obvious dependency problems which popped up due to some modules
  dependencies now being declared more precisely.
- Updating and centralizing the version of `rest-assured` which was
  partly specified incorrectly.

Note that while this patch fixes a lot of the Java 9 problems and nearly
all modules can now be built using Java 9, a few (unfortunately large)
problems still persist:

- The default date format for some languages has changed in Java 9 which
  causes a few tests in the admin interface to fail. I expect this to be
  harmless and think the tests just need adjustment, but I did not want
  to make that change without the ability to test it properly.
- We are on ElasticSearch version 1.7.6. Java 9 is supported starting
  with version 6.2. We had already discussed in the past that upgrading
  this is something we would like to do. Unfortunately, it's a lot of
  work.
- We also need to upgrade to Karaf 4.2.x. At the moment, the assembly
  steps fail due to us using an old Karaf version. Fortunately, we are
  already on 4.0.10 which isn't that far away (compared to ES). Still,
  it's some work.

Hence, consider this a first step. Obviously, I could also not test for
any runtime problems so far. The other problems need to be resolved
first.

Note though, that this does not only fix dependencies for Java 9. Most
of the problems are quite relevant for older Java versions as well.