MH-12813: Add audio and video track selection to video editor
This PR adds a new tab “Tracks” to the video editor that allows the
user to select which audio and video tracks will be published. In this
new tab, the user simply clicks on the video track (represented by a
thumbnail) or the audio track (represented by its waveform) to hide
it. Sanity checks are included, so its not possible to hide _every_
video and audio track there is, for example.

The information about which tracks are hidden is stored in asset
manager properties, which are then injected into the running
workflow(s). Thus, this PR depends on
[193](https://github.com/opencast/opencast/pull/193) (and is rebased
on it).

This _theoretically_ works for an arbitrary number of tracks in the
media package, but it currently is constrained to one and two tracks (the
most common use case). The allowed flavors flavors are configurable.

To make this possible, a specialized workflow operation handler
`SelectTracksWorkflowOperationHandler` had to be written, which is
capable of removing audio tracks from multiple input tracks, as well
as muxing a video track and an audio track from two different input
tracks (so it’s very similar to the `PrepareAV` WOH).

This work was sponsored by SWITCH.

Co-authored-by: Kim Rinnewitz <krinnewitz@plapadoo.de>
Co-authored-by: Sven Stauber <sven.stauber@switch.ch>