MH-12843, Fix “Add Event” Tab Index
This patch fixes the problem that hitting tab in the schedule form of
the “Add Event” wizard will not necessary bring you to the next input
field since the tab index of those fields is broken.

Note that the problem of jumping out of the form when reaching the end
still persists.