MH-12868: Remove unimplementable player adapter functionality
Given the current state of browsers' multimedia APIs,
frame-by-frame skipping and/or extracting the frame rate of a video
can not be implemented at the level it was envisioned here.
The frame rate would have to be passed in to the player adapter,
but I find that rather unsatisfying from an API design standpoint.
This functionality can much rather be implemented in the level above,
where we know more about the video than just its `HTMLMediaElement`,
i.e. the player directive.